+++ webchanges/mailer.py
@@ -52,7 +52,7 @@ class SMTPMailer(Mailer):
         self.auth = auth
         self.insecure_password = insecure_password
 
-    def send(self, msg: Union[EmailMessage]) -> None:
+    def send(self, msg: Optional[EmailMessage]) -> None:
         passwd = None
         if self.auth:
             if self.insecure_password:
@@ -70,8 +70,9 @@ class SMTPMailer(Mailer):
                 server.starttls()
             if self.auth:
                 server.login(self.smtp_user, passwd)
-            server.send_message(msg)
-            logger.info(f"SMTP email sent to {msg.get('to')} via {self.smtp_server}")
+            if msg:
+                server.send_message(msg)
+                logger.info(f"SMTP email sent to {msg.get('to')} via {self.smtp_server}")
 
 
 class SendmailMailer(Mailer):